issues with circular import for phasecalc
